feat(backend): container size feat

+++ src/main/java/ru/itmo/wastemanagement/dto/kioskorder/KioskOrderRowDto.java
@@ -16,7 +16,7 @@ public class KioskOrderRowDto {
     Integer garbagePointId;
     String garbagePointAddress;
 
-    Integer containerSizeId;
+    Long containerSizeId;
     String containerSizeCode;
 
     Integer fractionId;
